Added code to the Collision class
Hardcoded two RandomWalk objects in the constructor, used to test collisions

Added code to constructor to make the RandomWalk objects take a step. as long as there are more steps to be taken

Added get-methods to RandomWalk for X and Y coordinates.

Added a method 'samePosition' to test whether the two objects occupy the same space
	*'samePosition' takes two parameters, the two RandomWalk objects, and compare the X and Y coordinates between them. Returns true, if they occupy the same space

Added test code for the Collision objects to the Main class